Model loading with assimp
you would need to add assimp manually

Former-commit-id: a3bbbdc601d53aa2998763023faf540cea53f451

+++ Ignite-Core/include/Ignite/Renderer/IRendererAPI.h
@@ -28,7 +28,7 @@ namespace Ignite {
 
         virtual void Clear() = 0;
 
-        virtual void DrawIndexed(const IVertexBuffer* vertexBuffer, const IIndexBuffer* indexBuffer, uint16_t indexCount) = 0;
+        virtual void DrawIndexed(const IVertexBuffer* vertexBuffer, const IIndexBuffer* indexBuffer, uint32_t indexCount) = 0;
 
         virtual void SetUniformBufferObject(const UniformBufferObject& ubo) = 0;
